Run PR checks on Windows as well as Linux
We want to be able to support developers using Windows [[1]]. This
commit adds a build matrix for PR checks that includes running tests on
Windows as well as Linux, so that we can be more confident that any
future changes won't break Windows support.

Note that we also tweak the build matrix and the upload artifact
condition to only upload if running on Linux, to avoid uploading the
artifact twice. We're not sure what would happen if we didn't change
this, but we suspect that it wouldn't be good [[2]].

[1]: https://github.com/alphagov/govuk-design-system/issues/598
[2]: https://github.com/actions/upload-artifact#uploading-to-the-same-artifact